test(http2_adapter): rename authority test and drop unused dio_test dep
Agent-Logs-Url: https://github.com/cfug/dio/sessions/3e786e59-4495-4e7a-b0a4-37a262035db5
