+++ dio_test/pubspec.yaml
@@ -10,7 +10,6 @@ environment:
 
 dependencies:
   dio: any
-  http2: ^2.1.0
   path: any
 
 dev_dependencies:
